requirements.txt
pydub
numpy==1.23.4
Pillow==9.1.1
streamlit==1.26.0
PyYAML==6.0.1
pandas==1.4.4
ultralytics==8.0.175
ipython==8.5.0
Flask==2.3.3
# opencv-python==4.8.0
opencv-python-headless
python-multipart
pytube
uvicorn
fastapi
protobuf
